doc: reorg and cleanup
This is a significant update.

* The document is restructured to have a more logical flow.
* The LTS straw man is removed (it was just a straw man,
   after all, and it was causing confusion)
* The supported platforms bit is simplified to better reflect
   reality.
* Address breaking changes in the build
* Clarify the release tagging and maintenance branching model

The idea here is to simplify, clarify and define a branching model that
is closer to reality. It's a significant change that will require
additional review.

The LTS straw man is removed entirely. It can be added back in as a
separate document later on if necessary.